moves classes to their own files
